Trying to alleviate memory allocs during the run.  Made options for reference scanning and for 8-bit SSE alignment.

+++ aligner_cache.h
@@ -121,7 +121,7 @@ struct QKey {
 		s.resize(len);
 		uint64_t sq = seq;
 		for(int i = (len)-1; i >= 0; i--) {
-			s.set(sq & 3, i);
+			s.set((uint32_t)(sq & 3), i);
 			sq >>= 2;
 		}
 	}
